sending tx change alertdialog and progressdialog add ExchangeRates

+++ app/src/main/java/org/qtum/mromanovsky/qtum/ui/fragment/TransactionFragment/TransactionFragmentView.java
@@ -4,5 +4,5 @@ package org.qtum.mromanovsky.qtum.ui.fragment.TransactionFragment;
 import org.qtum.mromanovsky.qtum.ui.fragment.BaseFragment.BaseFragmentView;
 
 public interface TransactionFragmentView extends BaseFragmentView {
-    void setUpTransactionData(Long value, String receivedTime, String from, String to);
+    void setUpTransactionData(double value, String receivedTime, String from, String to);
 }
